+++ packages/google_sign_in/google_sign_in_ios/example/integration_test/google_sign_in_test.dart
@@ -1,0 +1,16 @@
+// Copyright 2013 The Flutter Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+import 'package:flutter_test/flutter_test.dart';
+import 'package:google_sign_in_platform_interface/google_sign_in_platform_interface.dart';
+import 'package:integration_test/integration_test.dart';
+
+void main() {
+  IntegrationTestWidgetsFlutterBinding.ensureInitialized();
+
+  testWidgets('Can initialize the plugin', (WidgetTester tester) async {
+    final GoogleSignInPlatform signIn = GoogleSignInPlatform.instance;
+    expect(signIn, isNotNull);
+  });
+}