+++ packages/webview_flutter/webview_flutter_wkwebview/ios/Classes/webview-umbrella.h
@@ -7,5 +7,6 @@
 #import <webview_flutter_wkwebview/FLTWKNavigationDelegate.h>
 #import <webview_flutter_wkwebview/FLTWKProgressionDelegate.h>
 #import <webview_flutter_wkwebview/FLTWebViewFlutterPlugin.h>
+#import <webview_flutter_wkwebview/FWFInstanceManager.h>
 #import <webview_flutter_wkwebview/FlutterWebView.h>
 #import <webview_flutter_wkwebview/JavaScriptChannelHandler.h>