+++ packages/webview_flutter/webview_flutter_wkwebview/lib/src/common/function_flutter_api_impls.dart
@@ -1,0 +1,25 @@
+// Copyright 2013 The Flutter Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+import 'instance_manager.dart';
+import 'web_kit.pigeon.dart';
+
+/// Flutter api to dispose functions.
+class FunctionFlutterApiImpl extends FunctionFlutterApi {
+  /// Constructs a [FunctionFlutterApiImpl].
+  FunctionFlutterApiImpl({InstanceManager? instanceManager}) {
+    this.instanceManager = instanceManager ?? InstanceManager.instance;
+  }
+
+  /// Maintains instances stored to communicate with native language objects.
+  late final InstanceManager instanceManager;
+
+  @override
+  void dispose(int instanceId) {
+    final Function? function = instanceManager.getInstance(instanceId);
+    if (function != null) {
+      instanceManager.removeInstance(function);
+    }
+  }
+}